# openapi_v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "EZ-Lang-FlashCard",
    "version": "1.0.0",
    "description": "REST API designed to provide flashcard functions for ez language",
    "contact": {
      "email": "vinhphuc.it@gmail.com",
      "name": "Phuc Tran"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/flashcard/"
    },
    {
      "url":"https://ez-lang-alpha.herokuapp.com/flashcard/"
    }
  ],
...
  "paths": {
...
    "/categories/{id}/flashCards":{
      "get": {
        "tags": [
          "Category"
        ],
        "summary": "Get FlashCards By Category",
        "operationId": "getFlashCardsByCategoryId",
        "parameters": [          
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },        
          {
            "name": "limit",
            "in": "query",
            "description": "maximum 100",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetFlashCardsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "GetFlashCardsResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseListResponse",
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FlashCard"
                }
              }
            }
          }
        ]
      },
...
      "BaseListResponse": {
        "type": "object",
        "properties": {
          "limit": {
            "type": "number"
          },
          "start": {
            "type": "number"
          },
          "totalItems": {
            "type": "number"
          },
          "sort": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "object"
            }
          }
        }
      },
      "FlashCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}